Bonus part done

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: elel-yak <elel-yak@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/24 16:17:20 by elel-yak          #+#    #+#             */
-/*   Updated: 2022/12/11 18:35:39 by elel-yak         ###   ########.fr       */
+/*   Updated: 2022/12/11 19:43:57 by elel-yak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@
 # include <unistd.h>
 # include <fcntl.h>
 
-int		ft_strlen(char *str);
+size_t	ft_strlen(char *str);
 char	*get_next_line(int fd);
 void	*ft_free(char *line);
